Provides user feedback on 404 errors and more
Added a catch statement to the AXIOS command that grabs a KIVA code and other info from the database to show the user when an address that might say Kansas City, MO, but might not be in that area. Also adjusted the appearance of the web page when it displays an address to be more uniform with it's appearance before an address is searched. Also, added some code to have cursor go back to a pointer if an address is not found.